+++ libs/core/src/lib/services/firmware-storage.service.interface.ts
@@ -5,4 +5,5 @@ export interface IFirmwareStorageService {
     contentType: string,
   ): Promise<void>;
   getSignedUrl(gcsPath: string): Promise<string>;
+  deleteObject(gcsPath: string): Promise<void>;
 }